pyproject.toml
[project]
name = "usmle-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langchain-pinecone>=0.2.8",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.2",
    "langsmith>=0.4.5",
    "pdfplumber>=0.11.7",
    "pinecone-client[grpc]>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "torch>=2.7.1",
    "transformers>=4.53.2",
]

[tool.uv.workspace]
members = [
    "venv",
]
